Added exception for todo not found. Api added to get one todo by id

+++ src/main/java/com/learnspring/demo/service/TodoService.java
@@ -1,10 +1,12 @@
 package com.learnspring.demo.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.learnspring.demo.exception.TodoNotFoundException;
 import com.learnspring.demo.model.Todo;
 import com.learnspring.demo.repository.TodoRepository;
 
@@ -13,15 +15,26 @@ public class TodoService {
     @Autowired
     private TodoRepository todoRepository;
 
-    public Todo createTodo(Todo todo){
+    public Todo createTodo(Todo todo) {
         return todoRepository.save(todo);
     }
 
-    public void deleteTodo(Long id){
+    public void deleteTodo(Long id) {
+        if (!todoRepository.existsById(id)) {
+            throw new TodoNotFoundException("Todo with id " + id + " not found");
+        }
         todoRepository.deleteById(id);
     }
 
-    public List<Todo> listTodos(){
+    public List<Todo> listTodos() {
         return todoRepository.findAll();
     }
+
+    public Todo getTodoById(Long id) {
+        if (!todoRepository.existsById(id)) {
+            throw new TodoNotFoundException("Todo with id " + id + " not found");
+        }
+        return todoRepository.findById(id).get();
+    }
+
 }
